refactor: separate getTimeAgo from Inquiry component
