+++ src/utils/getTimeAgo.js
@@ -1,0 +1,22 @@
+export default function getTimeAgo(createdAt) {
+  const now = new Date();
+  const timeDifference = now - new Date(createdAt);
+
+  const seconds = Math.floor(timeDifference / 1000);
+  const minutes = Math.floor(seconds / 60);
+  const hours = Math.floor(minutes / 60);
+  const days = Math.floor(hours / 24);
+  const weeks = Math.floor(days / 7);
+
+  if (weeks > 0) {
+    return `${weeks}주전`;
+  } else if (days > 0) {
+    return `${days}일전`;
+  } else if (hours > 0) {
+    return `${hours}시간전`;
+  } else if (minutes > 0) {
+    return `${minutes}분전`;
+  } else {
+    return `${seconds}초전`;
+  }
+}
